Fix migrations foreign _keys

+++ database/migrations/2017_10_17_170735_create_permission_user_table.php
@@ -13,7 +13,7 @@ class CreatePermissionUserTable extends Migration
         Schema::create($name, function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->unsignedBigInteger('permission_id')->index();
-            $table->unsignedBigInteger('user_id')->index();
+            $table->unsignedInteger('user_id')->index();
             
             $table->foreign('permission_id')->references('id')->on('permissions')->onDelete('cascade');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
