+++ database/migrations/2019_10_24_003025_create_clients_table.php
@@ -20,7 +20,7 @@ class CreateClientsTable extends Migration
             $table->string('surname', 128)->nullable();
             $table->string('address', 128)->nullable();
             $table->integer('user_id')->unsigned();
-            $table->integer('seller_id')->nullable()->unsigned();
+            $table->bigInteger('seller_id')->nullable()->unsigned();
             $table->enum('status', ['active', 'inactive'])->default('active');
 
             $table->foreign('user_id')
