+++ database/migrations/2021_02_10_233625_create_current_acounts_table.php
@@ -31,13 +31,13 @@ class CreateCurrentAcountsTable extends Migration
                 'pago_for_commissioner',
             ]);
             $table->decimal('pagandose', 8,2)->nullable();
-            $table->integer('client_id')->unsigned()->nullable();
+            $table->bigInteger('client_id')->unsigned()->nullable();
             $table->foreign('client_id')->references('id')->on('clients');
-            $table->integer('commissioner_id')->unsigned()->nullable();
+            $table->bigInteger('commissioner_id')->unsigned()->nullable();
             $table->foreign('commissioner_id')->references('id')->on('commissioners');
-            $table->integer('seller_id')->unsigned()->nullable();
+            $table->bigInteger('seller_id')->unsigned()->nullable();
             $table->foreign('seller_id')->references('id')->on('sellers');
-            $table->integer('sale_id')->unsigned()->nullable();
+            $table->bigInteger('sale_id')->unsigned()->nullable();
             $table->foreign('sale_id')->references('id')->on('sales');
             $table->timestamps();
         });
